move & rename dynamicBreadcrumb file
